codenav: Fix signed vs. unsigned comparisons

+++ codenav/src/codenavigation.c
@@ -120,7 +120,7 @@ static void load_configuration(void)
 	gchar *config_filename = NULL;
 	gchar **impl_list  = NULL, **head_list = NULL;
 	gsize head_list_len, impl_list_len;
-	int i;
+	gsize i;
 
 	/* Load user configuration */ 
 	config = g_key_file_new();
@@ -220,7 +220,7 @@ void plugin_cleanup(void)
 static void
 on_configure_response(GtkDialog* dialog, gint response, gpointer user_data)
 {
-	gint i=0;
+	gsize i=0;
 
 	GKeyFile *config = NULL;
 	gchar *config_filename = NULL;
